fix(test): split header-chrome tests under the file line limit
header_chrome/tests.rs sat at the 350-line cap after #756, so the next
assertion would fail File line limits. Split nav, announcement, and
page-chrome cases into sibling modules, matching html/tests.rs.
